fix: User routes failsafes

+++ src/users/users.controller.ts
@@ -7,17 +7,22 @@ import {
   ParseUUIDPipe,
   Patch,
   Post,
+  Req,
 } from '@nestjs/common';
+import { UserRole } from '@prisma/client';
+import { AuthedRequest } from 'src/auth/dto/sign-in.dto';
 import { Roles } from 'src/common/decorator/roles.decorator';
+import { isAdmin } from 'src/common/guards/roles.guard';
+import { UserCreateDto, UserUpdateDto } from './users.dto';
 import { UsersService } from './users.service';
-import { Prisma, UserRole } from '@prisma/client';
 
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
+  @Roles([UserRole.ADMIN])
   @Post()
-  create(@Body() data: Prisma.UserCreateInput) {
+  create(@Body() data: UserCreateDto) {
     return this.usersService.create(data);
   }
 
@@ -32,13 +37,33 @@ export class UsersController {
   }
 
   @Patch(':uuid')
-  update(@Param('uuid') uuid: string, @Body() data: Prisma.UserUpdateInput) {
+  update(
+    @Param('uuid') uuid: string,
+    @Body() data: UserUpdateDto,
+    @Req() request: AuthedRequest,
+  ) {
+    const admin = isAdmin(request.user);
+    // If the user is not an admin, they can only update their own user
+    if (uuid !== request.user.id && !admin) {
+      throw new Error('Unauthorized');
+    }
+
+    // Roles can only be updated by admins
+    if (data.roles && !admin) {
+      throw new Error('Unauthorized');
+    }
+
     return this.usersService.update(uuid, data);
   }
 
   @Delete(':uuid')
   @Roles([UserRole.ADMIN])
-  remove(@Param('uuid') uuid: string) {
+  remove(@Param('uuid') uuid: string, @Req() request: AuthedRequest) {
+    // Prevent admins from deleting themselves
+    if (uuid === request.user.id) {
+      throw new Error('Unauthorized');
+    }
+
     return this.usersService.remove(uuid);
   }
 }
